kdDebug() -> kdDebug(33001)
svn path=/trunk/koffice/; revision=161946

+++ kpresenter/kpresenter_doc.cc
@@ -3186,7 +3186,7 @@ void KPresenterDoc::insertFile(const QString & file )
     }
     KMacroCommand *macro = 0L;
     bool createMacro = false;
-    for ( int i = m_insertFilePage; i<m_pageList.count();i++)
+    for ( int i = m_insertFilePage; i<(int)m_pageList.count();i++)
     {
         if ( !macro )
             macro = new KMacroCommand( i18n("Insert File"));
